api: Move the actual node update logic out of node api
To keep API code straightforward, move the actual logic of NodeUpdate
and DiskUpdate from api/node.go to manager/node.go
